Rename and remove muon branches when reading
Renamed & removed the following:
- miniRelIso -> miniPFRelIso_all (common w/ electrons)
- chargedHadRelIso03 -> pfRelIso03_chg (common w/ electrons)
- miniIsoCharged -> miniPFRelIso_chg, compute by multiplying with lepton's pt, provided the getters (common w/ electrons)
- miniIsoNeutral -> computed from pt * (miniRelIso - miniRelIsoCharged), provided getters (common w/ electrons)
- mvaTTHjetNDauChargedMVASel -> REMOVED (not used in lepton selection anyways) (common w/ electrons)
- mvaTTHjetPtRel -> REMOVED (not used in lepton selection anyways) (common w/ electrons)
- jetBTagCSV -> jetBtagCSV (common w/ electrons)
- looseIdPOG -> REMOVED (set to constant = true in RecoMuon instances since the selected objects are always loose)
- mediumMuonId -> mediumId
- dpt_div_pt -> REMOVED (possible to calculate from ptErr, but not used anyways)
- segmentCompatibility -> segmentComp

+++ src/MuonHistManager.cc
@@ -49,8 +49,8 @@ void MuonHistManager::fillHistograms(const RecoMuon& muon, double evtWeight)
   fillWithOverFlow(histogram_jetPtRatio_, muon.jetPtRatio(), evtWeight, evtWeightErr);
   fillWithOverFlow(histogram_jetBtagCSV_, muon.jetBtagCSV(), evtWeight, evtWeightErr);
   fillWithOverFlow(histogram_tightCharge_, muon.tightCharge(), evtWeight, evtWeightErr);
-  fillWithOverFlow(histogram_passesLooseIdPOG_, muon.passesLooseIdPOG(), evtWeight, evtWeightErr);
-  fillWithOverFlow(histogram_passesMediumIdPOG_, muon.passesMediumIdPOG(), evtWeight, evtWeightErr);
+  fillWithOverFlow(histogram_passesLooseIdPOG_, static_cast<int>(muon.passesLooseIdPOG()), evtWeight, evtWeightErr);
+  fillWithOverFlow(histogram_passesMediumIdPOG_, static_cast<int>(muon.passesMediumIdPOG()), evtWeight, evtWeightErr);
   
   int abs_genPdgId = 0;
   if      ( muon.genLepton() ) abs_genPdgId = std::abs(muon.genLepton()->pdgId()); // generator level match to electron or muon
